fix: statistics transaction total type

+++ apps/api/domain/transaction_entity.go
@@ -54,7 +54,7 @@ type Transaction struct {
 
 type TransactionStatistic struct {
 	Date        string
-	Total       int32
+	Total       float32
 	TotalIncome float32
 }
